+++ src/main/java/ua/den/model/repository/BonuseRepository.java
@@ -3,5 +3,5 @@ package ua.den.model.repository;
 import ua.den.model.entity.dto.Bonuse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BonuseRepository extends JpaRepository<Bonuse, Integer> {
+public interface BonuseRepository extends JpaRepository<Bonuse, Long> {
 }
