+++ src/main/java/ua/den/model/repository/HarborRepository.java
@@ -3,5 +3,5 @@ package ua.den.model.repository;
 import ua.den.model.entity.dto.Harbor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HarborRepository extends JpaRepository<Harbor, Integer> {
+public interface HarborRepository extends JpaRepository<Harbor, Long> {
 }
