+++ src/main/java/ua/den/model/repository/RoomTypeRepository.java
@@ -3,5 +3,5 @@ package ua.den.model.repository;
 import ua.den.model.entity.dto.RoomType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomTypeRepository extends JpaRepository<RoomType, Integer> {
+public interface RoomTypeRepository extends JpaRepository<RoomType, Long> {
 }
