+++ src/main/java/ua/den/model/repository/ShipRepository.java
@@ -3,5 +3,5 @@ package ua.den.model.repository;
 import ua.den.model.entity.dto.Ship;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShipRepository extends JpaRepository<Ship, Integer> {
+public interface ShipRepository extends JpaRepository<Ship, Long> {
 }
