+++ src/main/java/ua/den/model/repository/TicketExcursionRepository.java
@@ -3,5 +3,5 @@ package ua.den.model.repository;
 import ua.den.model.entity.dto.TicketExcursion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketExcursionRepository extends JpaRepository<TicketExcursion, Integer> {
+public interface TicketExcursionRepository extends JpaRepository<TicketExcursion, Long> {
 }
